Fixes and docs (#176)
Changes:

- multiprocessing safe filehandling
- multi_tenancy with content_types
- schema compatibility of content types
- use the metaclass ability to register to registry
- proper copy of registry
- schema support in foreign keys
- register_default flag in tenant models meta
- deprecate registry attribute of fields
- use extended create_schema
- document parameters and example
- small renames related to ContentType
- automatically set via metaclasses a required ClassVar for
  no_constraint option of ContentType

+++ tests/contrib/multi_tenancy/test_mt_models.py
@@ -6,7 +6,7 @@ from uuid import UUID
 
 import pytest
 
-from edgy import Registry
+from edgy import Database, Registry
 from edgy.contrib.multi_tenancy import TenantModel
 from edgy.contrib.multi_tenancy.models import TenantMixin
 from edgy.core.db import fields
@@ -15,7 +15,7 @@ from edgy.testclient import DatabaseTestClient
 from tests.settings import DATABASE_URL
 
 database = DatabaseTestClient(DATABASE_URL)
-models = Registry(database=database)
+models = Registry(database=Database(database, force_rollback=True))
 
 pytestmark = pytest.mark.anyio
 
@@ -31,19 +31,17 @@ class StatusEnum(Enum):
 
 @pytest.fixture(autouse=True, scope="module")
 async def create_test_database():
-    try:
+    async with database:
         await models.create_all()
         yield
-        await models.drop_all()
-    except Exception as e:
-        pytest.skip(f"Error: {str(e)}")
+        if not database.drop:
+            await models.drop_all()
 
 
 @pytest.fixture(autouse=True)
 async def rollback_transactions():
-    with database.force_rollback():
-        async with database:
-            yield
+    async with models.database:
+        yield
 
 
 class Tenant(TenantMixin):
